move the scripts to phoenix package
